+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/kafkastreams/flushing/AsyncTransformer.scala
@@ -3,8 +3,7 @@ package com.twitter.finatra.kafkastreams.flushing
 import com.twitter.finagle.stats.{Stat, StatsReceiver, Verbosity}
 import com.twitter.finatra.kafkastreams.internal.utils.ProcessorContextLogging
 import com.twitter.util.{Duration, Future}
-import java.util.Collections
-import java.util
+import java.util.concurrent.ConcurrentHashMap
 import org.apache.kafka.streams.processor.internals.{
   InternalProcessorContext,
   ProcessorRecordContext
@@ -56,8 +55,8 @@ abstract class AsyncTransformer[K1, V1, K2, V2](
 
   @volatile private var flushOutputRecordsCancellable: Cancellable = _
 
-  val oldList = new util.LinkedList[(K2, V2, ProcessorRecordContext)]()
-  val outstandingResults = Collections.synchronizedList(oldList)
+  private val outstandingResults = ConcurrentHashMap
+    .newKeySet[(K2, V2, ProcessorRecordContext)](maxOutstandingFuturesPerTask)
 
   private val outstandingResultsGauge =
     statsReceiver.addGauge(Verbosity.Debug, "outstandingResults")(numOutstandingResults)